requirements.txt
altgraph==0.17.2
attrs==22.1.0
autopep8==1.7.0
colour==0.1.5
cycler==0.11.0
duckdb==0.4.0
et-xmlfile==1.1.0
fonttools==4.34.4
future==0.18.2
jsonschema==4.16.0
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pefile==2022.5.30
Pillow==9.2.0
pycodestyle==2.9.1
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
pyrsistent==0.18.1
PySide2==5.15.2.1
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
scipy==1.8.1
seaborn==0.11.2
shiboken2==5.15.2.1
six==1.16.0
styleframe==4.1
toml==0.10.2
